// src/common/enums/subsection.enum.ts
export enum Subsection {
  // Politics-related
  PresidentialOffice = 'presidential-office',
  Assembly = 'assembly',
  Diplomacy = 'diplomacy',
  Defense = 'defense',

  // Economy-related
  Finance = 'finance',
  RealEstate = 'real-estate',
  JobsStartup = 'jobs-startup',
  Consumer = 'consumer',

  // Industry-related
  Electronics = 'electronics',
  HeavyChem = 'heavy-chem',
  Automobile = 'automobile',
  Construction = 'construction',
  EnergyResources = 'energy-resources',
  TechScience = 'tech-science',
  Game = 'game',
  RetailService = 'retail-service',
  SmbVenture = 'smb-venture',
  BioHealth = 'bio-health',
  AgriLivestock = 'agri-livestock',
  MarineFishery = 'marine-fishery',

  // Local/Regions
  Gyeonggi = 'gyeonggi',
  Ulsan = 'ulsan',
  Gyeongnam = 'gyeongnam',
  DaeguGyeongbuk = 'daegu-gyeongbuk',
  GwangjuJeonnam = 'gwangju-jeonnam',
  Jeonbuk = 'jeonbuk',
  DaejeonChungnamSejong = 'daejeon-chungnam-sejong',
  Chungbuk = 'chungbuk',
  Gangwon = 'gangwon',
  Gwangju = 'gwangju',
  Busan = 'busan',
  Incheon = 'incheon',
  Jeju = 'jeju',

  // World-related
  USNA = 'us-na',
  China = 'cn',
  Japan = 'jp',
  AsiaAU = 'asia-au',
  EU = 'eu',
  LATAM = 'latam',
  MEA = 'mea',
  IGO = 'igo',

  // Society-related
  Incident = 'incident',
  CourtProsecution = 'court-prosecution',
  Education = 'education',
  WelfareLabor = 'welfare-labor',
  EnvironmentClimate = 'environment-climate',
  WomenChildren = 'women-children',
  OverseasKorean = 'overseas-korean',
  Multicultural = 'multicultural',

  // Culture-related
  BooksLit = 'books-lit',
  ComicsWebtoon = 'comics-webtoon',
  Religion = 'religion',
  PerformanceExhibition = 'performance-exhibition',
  AcademiaHeritage = 'academia-heritage',
  Media = 'media',
  TravelLeisure = 'travel-leisure',
  Life = 'life',

  // Entertainment
  Broadcast = 'broadcast',
  Movie = 'movie',
  KPop = 'kpop',
  GlobalEnt = 'global-ent',

  // Sports
  Baseball = 'baseball',
  Football = 'football',
  BasketballVolleyball = 'basketball-volleyball',
  Golf = 'golf',

  // Health
  HealthNote = 'health-note',
  WeeklyHealth = 'weekly-health',

  // Opinion
  Editorial = 'editorial',
  Column = 'column',
}